requirements.txt
customtkinter==5.2.2
moviepy==2.1.1
pillow==10.4.0
yt-dlp==2024.11.18
